Move estimator into own modules
